Archives/Prod/requirements.txt
boto3==1.14.0
scipy==1.2.1
numpy==1.16.2
yfinance==0.1.68
pandas==0.24.2
plotly==5.1.0
PyYAML==6.0
